refactor frombinarydata
